move the installer to the core bundle
