+++ core-bundle/src/Resources/config/migrations.yaml
@@ -7,6 +7,13 @@ services:
         arguments:
             - ~
 
+    contao.migration.command_compiler:
+        class: Contao\CoreBundle\Migration\CommandCompiler
+        public: true
+        arguments:
+            - '@database_connection'
+            - '@contao.doctrine.schema_provider'
+
     contao.migration.version_400.version_400_update:
         class: Contao\CoreBundle\Migration\Version400\Version400Update
         arguments:
